% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stae.R
\name{data_deal}
\alias{data_deal}
\title{data_deal}
\usage{
data_deal(bimr, aimr, bsd, asd, mg, pse)
}
\arguments{
\item{bimr}{before_iterative_mapping_result}

\item{aimr}{after_iterative_mapping_result}

\item{bsd}{before_sc_data}

\item{asd}{after_sc_data}

\item{mg}{marker_gene}

\item{pse}{pseudotime}
}
\description{
data_deal
}
