% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stae.R
\name{dataProcessing}
\alias{dataProcessing}
\title{dataProcessing}
\usage{
dataProcessing(bimr, aimr, bsd, asd, pse = 1)
}
\arguments{
\item{bimr}{before_iterative_mapping_result}

\item{aimr}{after_iterative_mapping_result}

\item{bsd}{before_sc_data}

\item{asd}{after_sc_data}

\item{pse}{pseudotime}
}
\description{
dataProcessing
}
