% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stae.R
\name{Parameter_settings}
\alias{Parameter_settings}
\title{parameter_setting
Hyperparameter setting}
\usage{
Parameter_settings(pdr, pseflag)
}
\arguments{
\item{pdr}{position_distance_ratio}

\item{pseflag}{Whether there is pseudo-temporal data(True/False) : True}

\item{dam}{Differentiation and migration of cell type}
}
\description{
Set Resolution and Cell Columns,resolution is the multiple of program amplification,
Cell Columns is the specified column name
}
